Remove false conditions from documentation
The statement "If sourceTime represents a date before midnight, January 1, 1970, gmtime returns NULL." is not true on any windows and visual studio versions I tested. If the input is -1, _gmtime32 returns corrupted structure instead of NULL, and _gmtime64 returns Thu Dec 31 23:59:59 1969.

The statement "_gmtime64, which uses the __time64_t structure, enables dates to be expressed up through 23:59:59, December 31, 3000, UTC" is also not true. For example it returns Mon Jan 19 08:00:00 3001 if the input is 32536800000.

The statement "_gmtime32 only represent dates through 23:59:59 January 18, 2038" is not true, since the maximum int value corresponds to 03:14:07 January 19, 2038.

The statement "Midnight, January 1, 1970, is the lower bound of the date range for both functions." is not true (see above).

The statement "If sourceTime is a null pointer, or if the sourceTime value is negative, these functions invoke an invalid parameter handler, as described in Parameter Validation. If execution is allowed to continue, the functions return NULL and set errno to EINVAL these functions invoke an invalid parameter handler, as described in Parameter Validation." is partially not true, since continuing of execution doesn't make the function return NULL both if the sourceTime is a null pointer or if the sourceTime is big negative value. If sourceTime is small negative value, the invalid parameter handle is not called.

Also, please consider patching these things in code instead of documentation.